--- src/actix/mod.rs
@@ -1,6 +1,0 @@
-mod api;
-mod auth;
-mod routes;
-mod start;
-
-pub use start::start_server;
