--- src/database/mod.rs
@@ -1,1 +1,0 @@
-pub mod init;
