+++ src/http/mod.rs
@@ -1,0 +1,8 @@
+mod auth;
+
+mod profile;
+
+mod routes;
+
+mod start;
+pub use start::start_http_server;
